requirements.txt
fastapi
uvicorn
redis
rq
python-dotenv
pydantic
groq
google-genai
sentence-transformers
scikit-learn
numpy
streamlit
requests